# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Public Collaborative Playlists
  version: 1.0.0
tags:
  - name: login
    description: Login
  - name: song
    description: Songs
  - name: request
    description: Song requests
  - name: admin
    description: Administrators
paths:
  /login:
    get:
      tags:
        - login
      summary: Log into your Spotify account
      operationId: login
      responses:
        302:
          description: Redirect to Spotify authorization page
  /callback:
    get:
      tags:
        - login
      summary: Get access and refresh tokens from Spotify authorization code and state
      operationId: callback
      parameters:
        - name: code
          in: query
          description: Authorization code
          schema:
            type: string
        - name: state
          in: query
          description: Authorization state
          schema:
            type: string
      responses:
        302:
          description: Redirect to application page
  /refresh_token:
    get:
      tags:
        - login
      summary: Get new access token from refresh token
      operationId: refreshToken
      parameters:
        - name: refresh_token
          in: query
          description: Refresh token
          schema:
            type: string
      responses:
        200:
          description: Obtained new access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
  /playlists/{playlistId}/requests:
    get:
      tags:
        - request
      summary: Get the song requests for a playlist
      operationId: getSongRequests
      parameters:
        - name: playlistId
          in: path
          description: ID of playlist
          required: true
          schema:
            type: string
      responses:
        200:
          description: Fetched song requests successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongRequestList'
    post:
      tags:
        - request
      summary: Request to add or remove songs to/from a playlist
      operationId: requestSongs
      parameters:
        - name: playlistId
          in: path
          description: ID of playlist
          required: true
          schema:
            type: string
      requestBody:
        description: Song IDs to request to add/remove
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SongUpdateList'
      responses:
        201:
          description: The requests for each of the songs have been created, or the requests for some of the songs have been created but the specified songs could not be added due to the state of the playlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistSongsDifference'
        422:
          description: One or more of the song IDs is not valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidSongList'
  /playlists/{playlistId}/requests/{requestId}/vote:
    post:
      tags:
        - request
      summary: Vote for a song request
      operationId: voteForRequest
      parameters:
        - name: playlistId
          in: path
          description: ID of playlist
          required: true
          schema:
            type: string
        - name: requestId
          in: path
          description: ID of request
          required: true
          schema:
            type: integer
      responses:
        200:
          description: The vote has been added
    delete:
      tags:
        - request
      summary: Remove vote for a song request
      operationId: removeVoteForRequest
      parameters:
        - name: playlistId
          in: path
          description: ID of playlist
          required: true
          schema:
            type: string
        - name: requestId
          in: path
          description: ID of request
          required: true
          schema:
            type: integer
      responses:
        200:
          description: The vote has been removed
  /playlists/{playlistId}/administrators:
    get:
      tags:
        - admin
      summary: Get the administrators of a playlist
      operationId: getAdministrators
      parameters:
        - name: playlistId
          in: path
          description: ID of playlist
          required: true
          schema:
            type: string
      responses:
        200:
          description: Fetched administrators successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
    put:
      tags:
        - admin
      summary: Update the administrators of a playlist
      operationId: updateAdministrators
      parameters:
        - name: playlistId
          in: path
          description: ID of playlist
          required: true
          schema:
            type: string
      requestBody:
        description: Administrator user IDs to add/remove
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateList'
      responses:
        200:
          description: The administrators have been updated
  /playlists/{playlistId}/requests/{requestId}:
    put:
      tags:
        - request
      summary: Approve or reject a request, or undo the acceptance or rejection of a request
      operationId: updateRequestStatus
      parameters:
        - name: playlistId
          in: path
          description: ID of playlist
          required: true
          schema:
            type: string
        - name: requestId
          in: path
          description: ID of request
          required: true
          schema:
            type: integer
      requestBody:
        description: New request status
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SongRequestStatusUpdate'
      responses:
        200:
          description: The status of the request has been updated. If "approved" or "rejected", the request will be removed from the database at the given timestamp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionTimestamp'
        404:
          description: The request or playlist does not exist
  /playlists/{playlistId}/songs:
    get:
      tags:
        - song
      summary: Get the songs in a playlist
      operationId: getSongs
      parameters:
        - name: playlistId
          in: path
          description: ID of playlist
          required: true
          schema:
            type: string
      responses:
        200:
          description: Fetched songs successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SongList'
    put:
      tags:
        - song
      summary: Update the songs in a playlist
      operationId: updateSongs
      parameters:
        - name: playlistId
          in: path
          description: ID of playlist
          required: true
          schema:
            type: string
      requestBody:
        description: Song IDs to add/remove
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SongUpdateList'
      responses:
        200:
          description: The songs have been updated
        409:
          description: One or more of the specified songs conflict with the playlist state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistSongsDifference'
components:
  schemas:
    AccessToken:
      type: object
      properties:
        access_token:
          type: string
    SongList:
      type: object
      properties:
        areYouAdmin:
          type: boolean
        songs:
          type: array
          items:
            $ref: '#/components/schemas/Song'
    Song:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        dateAdded:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in milliseconds
    SongRequestList:
      type: object
      properties:
        areYouAdmin:
          type: boolean
        addRequests:
          type: array
          items:
            $ref: '#/components/schemas/SongRequest'
        removeRequests:
          type: array
          items:
            $ref: '#/components/schemas/SongRequest'
    SongRequest:
      type: object
      properties:
        requestId:
          type: integer
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        dateAdded:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in milliseconds
        numVotes:
          type: integer
        hasYourVote:
          type: boolean
    SongUpdateList:
      type: object
      properties:
        songsToAdd:
          type: array
          items:
            type: string
        songsToRemove:
          type: array
          items:
            type: string
    PlaylistSongsDifference:
      type: object
      properties:
        songsAlreadyInPlaylist:
          type: array
          items:
            type: string
        songsNotInPlaylist:
          type: array
          items:
            type: string
    InvalidSongList:
      type: object
      properties:
        invalidSongIds:
          type: array
          items:
            type: string
    UserUpdateList:
      type: object
      properties:
        usersToAdd:
          type: array
          items:
            type: string
        usersToRemove:
          type: array
          items:
            type: string
    UserList:
      type: object
      properties:
        users:
          type: array
          items:
            type: string
    DeletionTimestamp:
      type: object
      properties:
        deleteAt:
          type: string
          format: date-time
    SongRequestStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]